3.17 수업

+++ MyFirstWeb/src/main/java/kr/co/ict/servlet/service/BoardDeleteService.java
@@ -2,6 +2,7 @@ package kr.co.ict.servlet.service;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import kr.co.ict.BoardDAO;
 
@@ -11,10 +12,20 @@ public class BoardDeleteService implements IBoardService{
 	public void excute(HttpServletRequest request, HttpServletResponse response) {
 		String boardNum = request.getParameter("board_num");
 		int bNum = Integer.parseInt(boardNum);
+		 
+	
+		// 아이디 체킹 1. 폼에서 날린 writer에 든 값이랑 세션 내 아이디값 변수로 저장
+		String writer = request.getParameter("writer");
 		
-		// dao 생성
-		BoardDAO dao = BoardDAO.getInstance();
-		// 삭제로직 실행
-		dao.deleteBoard(bNum);		
+		HttpSession session = request.getSession();
+		String sId = (String)session.getAttribute("session_id");
+		
+		// 아래 DAO생성 및 삭제로직을 if문을 이용해 감싸주고, if문은 글쓴이가 현재 로그인한 사람이 맞는지 체크		
+		if(sId.equals(writer)) {
+			// dao 생성
+			BoardDAO dao = BoardDAO.getInstance();
+			// 삭제로직 실행
+			dao.deleteBoard(bNum);		
+		}
 	}
 }
